Moved ISS-related code to specific Spark Helper
