+++ java/code/src/com/suse/manager/iss/RouteWithIssToken.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 SUSE LLC
+ * Copyright (c) 2024--2025 SUSE LLC
  *
  * This software is licensed to you under the GNU General Public License,
  * version 2 (GPLv2). There is NO WARRANTY for this software, express or
@@ -8,7 +8,7 @@
  * along with this software; if not, see
  * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
  */
-package com.suse.manager.webui.utils;
+package com.suse.manager.iss;
 
 import com.suse.manager.webui.utils.token.Token;
 
@@ -18,7 +18,8 @@ import spark.Response;
 /**
  * A route that gets the authentication token in addition to the request and response.
  */
-public interface RouteWithToken {
+@FunctionalInterface
+public interface RouteWithIssToken {
 
     /**
      * Invoked when a request is made on this route's corresponding path.
@@ -26,7 +27,8 @@ public interface RouteWithToken {
      * @param request the request object
      * @param response the response object
      * @param token the token with this request
+     * @param serverFqdn the FQDN of the remote server
      * @return the content to be set in the response
      */
-    Object handle(Request request, Response response, Token token);
+    Object handle(Request request, Response response, Token token, String serverFqdn);
 }
